pyproject.toml
[project]
name = "mistake-notebook"
version = "1.0.0"
description = "错题库桌面应用程序"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.1",
]

[project.scripts]
mistake-notebook = "mistake_notebook.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mistake_notebook"]

[dependency-groups]
dev = [
    "pyinstaller>=6.0",
]
